Simplify debian install script
set -e exits the script if something fails.
Prefixing all root operations with sudo makes it work by default on Ubuntu and debian setups with sudo (recommended).
phpmyadmin is not necessary.
node shouldn't be compiled as root, and xdg-open should not be run as root.
Use -y flag to apt-get to cut down messages.
Create tmp directory automatically (it's not checked in to git).
Less fluff in the output.